CRAYSAT-1944: Improve performance of ProductCatalog
Improve the performance of loading the data from the product catalog
Kubernetes ConfigMaps during construction of `ProductCatalog` as
follows:

* Only load and validate the product catalog schema file against the
  JSON Schema metaschema once and re-use the validator across all
  instances of InstalledProductVersion.
* Add a `shallow` boolean optional argument to `ProductCatalog` that
  allows it to only load product catalog data from the main
  `cray-product-catalog` ConfigMap if that's all that the user of the
  library needs.
* Cache the value of the `is_valid` property on the
  `InstalledProductVersion`. This only needs to be computed once. It is
  accessed twice in the `ProductCatalog` constructor.

Fix existing unit tests to work with the refactored code. The tests now
mock only the Kubernetes API calls, so we are testing all the code
called by the constructor of the `ProductCatalog`. Add tests for the
`shallow` parameter that was added to `ProductCatalog`. Also add tests
for `get_product` when the requested product does not exist in the
product catalog.

Test Description:
Unit tests pass.

Included the new version in a build of the `cray-sat` container and
tested on rocket.